Moved file functions to files.c
Ticket: None
Changelog: None
