+++ lib/block.c
@@ -6,6 +6,7 @@
 #include <time.h>
 
 #include "definitions.h"
+#include "files.h"
 #include "head.h"
 #include "logger.h"
 #include "utils.h"
