+++ lib/files.c
@@ -1,0 +1,111 @@
+#include "files.h"
+
+#include <assert.h>
+#include <errno.h>
+#include <memory.h>
+#include <stdarg.h>
+#include <string.h>
+#include <sys/stat.h>
+#include <unistd.h>
+
+#include "definitions.h"
+#include "logger.h"
+
+/******************************************************************************/
+
+bool LCH_FileSize(FILE *file, size_t *size) {
+  if (fseek(file, 0, SEEK_END) != 0) {
+    LCH_LOG_ERROR("Failed to seek to end of file: %s", strerror(errno));
+    return false;
+  }
+
+  long pos = ftell(file);
+  if (pos < 0) {
+    LCH_LOG_ERROR("Failed to obtain the current file position indicator: %s",
+                  strerror(errno));
+    return false;
+  }
+  *size = (size_t)pos;
+
+  if (fseek(file, 0, SEEK_SET) != 0) {
+    LCH_LOG_ERROR("Failed to seek to start of file: %s", strerror(errno));
+    return false;
+  }
+
+  return true;
+}
+
+/******************************************************************************/
+
+bool LCH_FileExists(const char *const path) {
+  struct stat sb;
+  memset(&sb, 0, sizeof(struct stat));
+  return stat(path, &sb) == 0;
+}
+
+/******************************************************************************/
+
+bool LCH_IsRegularFile(const char *const path) {
+  struct stat sb;
+  memset(&sb, 0, sizeof(struct stat));
+  return (stat(path, &sb) == 0) && ((sb.st_mode & S_IFMT) == S_IFREG);
+}
+
+/******************************************************************************/
+
+bool LCH_IsDirectory(const char *const path) {
+  struct stat sb;
+  memset(&sb, 0, sizeof(struct stat));
+  return (stat(path, &sb) == 0) && ((sb.st_mode & S_IFMT) == S_IFDIR);
+}
+
+/******************************************************************************/
+
+bool LCH_PathJoin(char *path, const size_t path_max, const size_t n_items,
+                  ...) {
+  assert(path_max >= 1);
+
+  va_list ap;
+  va_start(ap, n_items);
+
+  size_t used = 0;
+  bool truncated = false;
+  for (size_t i = 0; i < n_items; i++) {
+    if (i > 0) {
+      if (path_max - used < 2) {
+        truncated = true;
+        break;
+      }
+      path[used++] = LCH_PATH_SEP;
+    }
+
+    char *const sub = va_arg(ap, char *);
+    const size_t sub_len = strlen(sub);
+    for (size_t j = 0; j < sub_len; j++) {
+      if (path_max - used < 2) {
+        truncated = true;
+        break;
+      }
+      path[used++] = sub[j];
+    }
+  }
+
+  va_end(ap);
+  path[used] = '\0';
+
+  if (truncated) {
+    LCH_LOG_ERROR("Failed to join paths: Truncation error.");
+    return false;
+  }
+  return true;
+}
+
+/******************************************************************************/
+
+bool LCH_FileDelete(const char *const filename) {
+  if (unlink(filename) != 0) {
+    LCH_LOG_ERROR("Failed to delete file '%s': %s", filename, strerror(errno));
+    return false;
+  }
+  return true;
+}
