+++ lib/utils.c
@@ -140,152 +140,6 @@ char *LCH_StringStrip(char *str, const char *charset) {
 
 /******************************************************************************/
 
-bool LCH_FileSize(FILE *file, size_t *size) {
-  if (fseek(file, 0, SEEK_END) != 0) {
-    LCH_LOG_ERROR("Failed to seek to end of file: %s", strerror(errno));
-    return false;
-  }
-
-  long pos = ftell(file);
-  if (pos < 0) {
-    LCH_LOG_ERROR("Failed to obtain the current file position indicator: %s",
-                  strerror(errno));
-    return false;
-  }
-  *size = (size_t)pos;
-
-  if (fseek(file, 0, SEEK_SET) != 0) {
-    LCH_LOG_ERROR("Failed to seek to start of file: %s", strerror(errno));
-    return false;
-  }
-
-  return true;
-}
-
-/******************************************************************************/
-
-bool LCH_FileExists(const char *const path) {
-  struct stat sb;
-  memset(&sb, 0, sizeof(struct stat));
-  return stat(path, &sb) == 0;
-}
-
-/******************************************************************************/
-
-bool LCH_IsRegularFile(const char *const path) {
-  struct stat sb;
-  memset(&sb, 0, sizeof(struct stat));
-  return (stat(path, &sb) == 0) && ((sb.st_mode & S_IFMT) == S_IFREG);
-}
-
-/******************************************************************************/
-
-bool LCH_IsDirectory(const char *const path) {
-  struct stat sb;
-  memset(&sb, 0, sizeof(struct stat));
-  return (stat(path, &sb) == 0) && ((sb.st_mode & S_IFMT) == S_IFDIR);
-}
-
-/******************************************************************************/
-
-bool LCH_PathJoin(char *path, const size_t path_max, const size_t n_items,
-                  ...) {
-  assert(path_max >= 1);
-
-  va_list ap;
-  va_start(ap, n_items);
-
-  size_t used = 0;
-  bool truncated = false;
-  for (size_t i = 0; i < n_items; i++) {
-    if (i > 0) {
-      if (path_max - used < 2) {
-        truncated = true;
-        break;
-      }
-      path[used++] = LCH_PATH_SEP;
-    }
-
-    char *const sub = va_arg(ap, char *);
-    const size_t sub_len = strlen(sub);
-    for (size_t j = 0; j < sub_len; j++) {
-      if (path_max - used < 2) {
-        truncated = true;
-        break;
-      }
-      path[used++] = sub[j];
-    }
-  }
-
-  va_end(ap);
-  path[used] = '\0';
-
-  if (truncated) {
-    LCH_LOG_ERROR("Failed to join paths: Truncation error.");
-    return false;
-  }
-  return true;
-}
-
-/******************************************************************************/
-
-char *LCH_FileRead(const char *const path, size_t *const length) {
-  LCH_Buffer *const buffer = LCH_BufferCreate();
-  if (buffer == NULL) {
-    return NULL;
-  }
-
-  if (!LCH_BufferReadFile(buffer, path)) {
-    LCH_BufferDestroy(buffer);
-    return NULL;
-  }
-
-  if (length != NULL) {
-    *length = LCH_BufferLength(buffer);
-  }
-
-  char *str = LCH_BufferToString(buffer);
-  assert(str != NULL);
-  return str;
-}
-
-/******************************************************************************/
-
-bool LCH_FileWrite(const char *const path, const char *const str) {
-  if (!LCH_CreateParentDirectories(path)) {
-    LCH_LOG_ERROR("Failed to create parent directories for file '%s'", path);
-    return false;
-  }
-
-  FILE *file = fopen(path, "w");
-  if (file == NULL) {
-    LCH_LOG_ERROR("Failed to open file '%s' for writing: %s", path,
-                  strerror(errno));
-    return false;
-  }
-
-  if (fputs(str, file) == EOF) {
-    LCH_LOG_ERROR("Failed to write to file '%s'.", path);
-    fclose(file);
-    return false;
-  }
-
-  fclose(file);
-  return true;
-}
-
-/******************************************************************************/
-
-bool LCH_FileDelete(const char *const filename) {
-  if (unlink(filename) != 0) {
-    LCH_LOG_ERROR("Failed to delete file '%s': %s", filename, strerror(errno));
-    return false;
-  }
-  return true;
-}
-
-/******************************************************************************/
-
 static bool IndicesOfFieldsInHeader(size_t *const indices,
                                     const LCH_List *const fields,
                                     const LCH_List *const header) {
