+++ lib/utils.h
@@ -33,43 +33,6 @@ bool LCH_StringStartsWith(const char *str, const char *substr);
 
 char *LCH_StringStrip(char *str, const char *charset);
 
-bool LCH_FileSize(FILE *file, size_t *size);
-
-bool LCH_FileExists(const char *path);
-
-bool LCH_IsRegularFile(const char *path);
-
-bool LCH_IsDirectory(const char *path);
-
-bool LCH_PathJoin(char *path, size_t path_max, size_t n_items, ...);
-
-/**
- * @brief Read a text file.
- *
- * @note Returned string must be freed with free(3).
- * @param[in] path path to file.
- * @param[out] length If not NULL, it is set to be the string length, excluding
- *                    the terminating null byte.
- * @return string containing the file content, or NULL in case of error.
- */
-char *LCH_FileRead(const char *path, size_t *length);
-
-/**
- * @brief Write a text file.
- *
- * @param[in] path path to file.
- * @param[in] str string to write to file.
- * @return false in case of error.
- */
-bool LCH_FileWrite(const char *path, const char *str);
-
-/**
- * @brief Delete file.
- * @param filename File to delte.
- * @return True on success, otherwise false.
- */
-bool LCH_FileDelete(const char *filename);
-
 LCH_Json *LCH_TableToJsonObject(const LCH_List *table,
                                 const LCH_List *primary_fields,
                                 const LCH_List *subsidiary_fields);
@@ -85,8 +48,6 @@ bool LCH_ParseVersion(const char *str, size_t *major, size_t *minor,
 char *LCH_StringDuplicate(const char *str);
 
 char *LCH_StringFormat(const char *format, ...);
-
-bool LCH_CreateParentDirectories(const char *filename);
 
 bool LCH_ListInsertBufferDuplicate(LCH_List *list, size_t index,
                                    const LCH_Buffer *buffer);
